[Feature] Make the codebase compatible with Python 3.8 and newer (#173)

+++ torchdr/utils/optim.py
@@ -8,7 +8,7 @@
 import torch
 from tqdm import tqdm
 
-DTYPE = torch.double
+DTYPE = torch.float32
 DEVICE = "cpu"
 
 OPTIMIZERS = {
